Separate modal from session form in components
